added search filters for /parent-management/parents/get-organization-tickets
